- improvements in GUI
- added the fine-grain authorization protection layer into the web
  application (both UI and PHP backend)

- fixed bugs.

M    doc/ChangeLog
M    web/dynamic/ShowAttachment.php
M    web/dynamic/DisplayRun.php
M    web/dynamic/SearchFormParams.php
D    web/dynamic/RequestFFEntries.php
M    web/dynamic/DisplayRunParams.php
M    web/dynamic/DeleteFFEntry.php
D    web/dynamic/Menubar_saved.js
M    web/dynamic/RequestRuns.php
M    web/dynamic/CreateShift.php
M    web/dynamic/CloseShift.php
M    web/dynamic/DisplayFFEntry.php
M    web/dynamic/DisplayShift.php
M    web/dynamic/RequestInfo.php
M    web/dynamic/Search.php
M    web/dynamic/DisplayExperiment.php
M    web/dynamic/index.php
M    web/dynamic/Dialogs.js
M    web/dynamic/RequestTags.php
M    web/dynamic/RequestShifts.php
M    web/dynamic/RequestExperiments.php
M    web/dynamic/UpdateFFEntry.php
M    web/dynamic/RequestAttachments.php
M    web/dynamic/NewFFEntry.php
M    web/TODO
AM   web/LogBookAuth.class.php
M    web/LogBookFFAttachment.class.php


git-svn-id: https://pswww.slac.stanford.edu/svn-readonly/psdmrepo@539 b967ad99-d558-0410-b138-e0f6c56caec7

+++ LogBook/trunk/web/dynamic/DeleteFFEntry.php
@@ -5,8 +5,6 @@ require_once('LogBook/LogBook.inc.php');
 /*
  * This script will process a request for deleting the specified free-form entry.
  */
-if( !LogBookAuth::isAuthenticated()) return;
-
 if( isset( $_POST['id'] )) {
     $id = trim( $_POST['id'] );
     if( $id == '' )
@@ -33,9 +31,24 @@ try {
         or die( "no such free-form entry" );
 
     $experiment = $entry->parent();
+    $instrument = $experiment->instrument();
 
+    // Check for the authorization
+    //
+    if( !LogBookAuth::instance()->canDeleteMessages( $experiment->id())) {
+        print( LogBookAuth::reporErrorHtml(
+            'You are not authorized to delete messages of the experiment',
+            'index.php?action=select_experiment'.
+                '&instr_id='.$instrument->id().
+                '&instr_name='.$instrument->name().
+                '&exper_id='.$experiment->id().
+                '&exper_name='.$experiment->name()));
+        exit;
+    }
+
+    // Proceed to the operation
+    //
     $logbook->delete_entry_by_id( $id );
-    $logbook->commit();
 
     // Return back to the caller
     //
@@ -67,6 +80,7 @@ try {
             ;
         }
     }
+    $logbook->commit();
 
 } catch( RegDBException $e ) {
     print $e->toHtml();
